Migrate MinIO -> Garage store for objects (#112)
- Added basic 1-node Garage configuration

- Switched librarian to use Garage

This carries a problem that Milvus is reliant on MinIO, but the Milvus configuration is currently broken anyway.

This also adds a Ceph configuration which is broken - started getting Ceph to work but proving very difficult to do a Ceph deployment for the general case, limitations on how Ceph fixes IP addresses in the monmap which won't generalise to a general K8s / docker compose configurator.